Split show/show_list stuff from ci_processing.
